requirement.txt
Django==5.0.2
psycopg2-binary==2.9.9
requests==2.31.0
Pillow==10.2.0
django-rest-framework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
python-dotenv==1.0.1

# Additional recommended packages for a full book-lending website
Django==5.0.2
psycopg2-binary==2.9.9
requests==2.31.0
Pillow==10.2.0
django-rest-framework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
python-dotenv==1.0.1
whitenoise==6.6.0
celery==5.3.6
redis==5.0.1
django-crispy-forms==2.1
django-allauth==0.57.2
crispy-bootstrap5==0.7
django-environ==0.11.2
django-debug-toolbar==4.3.0
drf-yasg==1.21.7
uvicorn==0.29.0
gunicorn==21.2.0
boto3==1.34.11